# fix(amazonq): Reduce plugin start-up latency